Refactored #1, created an abstract class and used variable types with Generics
