+++ stats/src/CsvFileReader.ts
@@ -1,12 +1,12 @@
 import fs from 'fs';
-import { dateStringToDate } from './utils';
+
 import { MatchResult } from './MatchResult';
 
-type MatchData = [Date, string, string, number, number, MatchResult, string];
-
-export class CsvFileReader {
-  data: MatchData[] = [];
+//Using Generics to leave the datype open as a parameter. By convenction you use T for TypeOfData parameter
+export abstract class CsvFileReader<T> {
+  data: T[] = [];
   constructor(public filename: string) {}
+  abstract mapRow(row: string[]): T;
   read(): void {
     this.data = fs
       .readFileSync(this.filename, {
@@ -16,18 +16,6 @@ export class CsvFileReader {
       .map((row: string): string[] => {
         return row.split(',');
       })
-      .map(
-        (row: string[]): MatchData => {
-          return [
-            dateStringToDate(row[0]),
-            row[1],
-            row[2],
-            parseInt(row[3]),
-            parseInt(row[4]),
-            row[5] as MatchResult, //"H", "A", "D"
-            row[6],
-          ];
-        }
-      );
+      .map(this.mapRow);
   }
 }
