+++ stats/src/MatchReader.ts
@@ -1,0 +1,18 @@
+import { CsvFileReader } from './CsvFileReader';
+import { dateStringToDate } from './utils';
+import { MatchResult } from './MatchResult';
+
+type MatchData = [Date, string, string, number, number, MatchResult, string];
+export class MatchReader extends CsvFileReader<MatchData> {
+  mapRow(row: string[]): MatchData {
+    return [
+      dateStringToDate(row[0]),
+      row[1],
+      row[2],
+      parseInt(row[3]),
+      parseInt(row[4]),
+      row[5] as MatchResult, //"H", "A", "D"
+      row[6],
+    ];
+  }
+}
