+++ stats/src/index.ts
@@ -1,8 +1,8 @@
 //import file system from node modules. To stop the error, load type definition file
-import { CsvFileReader } from './CsvFileReader';
+import { MatchReader } from './MatchReader';
 import { MatchResult } from './MatchResult';
 //Parse the data
-const reader = new CsvFileReader('football.csv');
+const reader = new MatchReader('football.csv');
 reader.read();
 
 const dateOfFirstMatch = reader.data[0][0];
